pyproject.toml
[project]
name = "image-dub"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "faiss-cpu>=1.11.0.post1",
    "matplotlib>=3.10.3",
    "open-clip-torch>=3.0.0",
    "opencv-python>=4.12.0.88",
    "pytesseract>=0.3.13",
    "tqdm>=4.67.1",
    "transformers>=4.53.3",
]
